modules: Simplify modules
Drop the idea that there's dynamic modules, and focus on working with
the expected sections that modules would care about. It's likely a
special section/symbol might be useful in the future for exporting
metadata and symbol tables but for MVP hello world type stuff all
that should be needed is loading relevant sections, applying
relocations (linking!) for things like other section addresses
and a printk function call.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>