modules: Print out lots of relocation info
Adds a global symbol table with one symbol (printk). Finding
symbols in the future probably should be overridden so a perfect hash or
binary search can be used. But... that's for the future.

For now we can find the printk symbol in the global symbol table, and
see (I believe) the associated op code entry in the .text section that
needs it.

One more step towards linking I hope.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>