modules: Mostly working linking with sections
For branch-link (call) instructions for far away targets a
veneer/trampoline function needs to be generated by the linker.

This is complex as now some intermediate code needs to be placed in
the loadable module .text section somewhere possibly requiring
re-allocation of the section and such.

It's unclear to me at this moment what the best route to deal with this.

But the general problem is ARM32_R_CALL relocation on a BL instruction
where the instruction address is in SRAM and the target is in ROM
may be very very far apart in the memory space. BL expects a PC relative
signed offset to branch-link to which is impossible to directly do.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>