drivers/regulator: add support for the BQ25120A PMIC
Add support for the BQ25120A PMIC. This IC is a highly integrated battery
charge management solution for wearables and IoT, including both a
charger and 2 voltage regulators (one buck and one LDO).

This commit adds support for the regulator part of the IC. The driver
supports setting the voltage and enabling/disabling the regulators.

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>