llext: [1/2] add len parameter to llext_peek
llext_peek is used to obtain a pointer to a position in the buffer, if
the loader supports this functionality. However, the original interface
did not allow specifying the length of the data that would be peeked,
thus making it impossible to use this function in a safe way.

This patch adds a len parameter to llext_peek, which allows the function
to check if the requested data is within the bounds of the buffer before
returning a pointer to it.

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>