cmake: (refact) drop redundant code
The 'convert_list_of_flags_to_string_of_flags' function was added back
in 2017 with commit 89516fbc25f0 at the time language support was added
to the CMake build system. The same commit also added all users of that
function; they were directly passed the result of get_property() calls,
so a simple replace was sufficient.

This code was later expanded with the use of generator expressions to
manage delimiters and prefixes, which required replacing the ';'
separator with '$<SEMICOLON>' to properly preserve the list entries.

This made the conversion function invalid, since no ';' can be found in
any 'zephyr_get_*' output, so the conversion function was effectively a
no-op; setting a space as a delimiter is already sufficient to get the
desired result since circa 2020.

Remove this function and 'get_property_and_add_prefix', a macro that was
made redundant by the same value-to-genex conversion.

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>