llext-edk: collect information from app target
Sometimes the data in 'zephyr_interface' is not enough to build
extensions that use certain Zephyr modules. For example, 'fatfs'
defines extra include directories and compile definitions in its own
interface library, which is then linked to the 'app' target.

Getting the compile definitions and include directories from the 'app'
target includes those modules as well.

Signed-off-by: Luca Burelli <l.burelli@arduino.cc>